requirements.txt
# 生成环境依赖
colorlog==6.8.2
pillow==11.0.0
anyio==4.6.2.post1 
certifi==2024.8.30 
exceptiongroup==1.2.2 
h11==0.14.0 
httpcore==1.0.7 
httpx==0.28.0 
idna==3.10 
sniffio==1.3.1 
typing-extensions==4.12.2
annotated-types==0.7.0
pydantic==2.10.3
pydantic-core==2.27.1
pyyaml==6.0.2